Moved state up to the page to keep component stateless, better for passing down state to the supply
